main_api/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.10.4
amqp==5.1.1
anyio==3.6.1
argcomplete==1.10.3
async-timeout==4.0.2
attrs==22.1.0
azure.storage.blob==12.17.0
beautifulsoup4==4.8.2
billiard==3.6.4.0
blis==0.7.8
boto3==1.26.13
botocore==1.29.13
bpemb==0.3.4
cachetools==5.2.0
catalogue==2.0.8
celery==5.2.7
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
compressed-rtf==1.0.6
confection==0.0.2
conllu==4.5.2
contourpy==1.0.5
cycler==0.11.0
cymem==2.0.6
datasets==2.11.0
date-parser==0.3
dateparser==1.1.8
Deprecated==1.2.13
dill==0.3.6
docx2txt==0.8
ebcdic==1.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0-py3-none-any.whl
extract-msg==0.28.7
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
filelock==3.8.0
flair==0.12.2
flower==1.2.0
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
httplib2==0.20.4
huggingface-hub==0.13.4
humanize==4.6.0
hyperopt==0.2.7
idna==3.4
IMAPClient==2.1.0
importlib-metadata==3.10.1
install==1.3.5
Janome==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.1
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.0
more-itertools==8.14.0
mpld3==0.3
multidict==6.0.2
multiprocess==0.70.14
murmurhash==1.0.8
networkx==2.8.7
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.3
olefile==0.46
openai==0.27.2
overrides==3.1.0
packaging==21.3
pandas==1.4.4
pathy==0.6.2
pdf2image==1.16.0
pdfminer.six==20191110
phonenumbers==8.12.56
Pillow==9.2.0
pptree==3.1
preshed==3.0.7
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.1
psycopg2-binary==2.9.3
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pydantic==1.9.2
PyMuPDF==1.19.6
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-multipart==0.0.5
python-pptx==0.6.21
pytorch_revgrad==0.2.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
regex==2022.3.2
requests==2.28.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
segtok==1.5.11
sentencepiece==0.1.97
sentry-sdk==1.9.9
six==1.12.0
slackclient==2.9.4
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-alignments==0.8.5
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
SpeechRecognition==3.8.1
SQLAlchemy==1.4.41
sqlitedict==2.0.0
srsly==2.4.4
starlette==0.20.4
tabulate==0.8.10
textract==1.6.5
thinc==8.1.2
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
torch==1.11.0
tornado==6.3.1
tqdm==4.64.1
transformer-smaller-training-vocab==0.2.3
transformers==4.21.3
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.5
Wikipedia-API==0.5.4
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.0.3
xxhash==3.2.0
yarl==1.8.1
zipp==3.8.1
